consumer average rps in app.log

+++ fraud-detection-service/src/main/java/com/example/fraud_detection_service/adapter/KafkaClient.java
@@ -54,7 +54,7 @@ public class KafkaClient {
 
         if (startTime.get() != null) {
             double ms = (endTime - startTime.get()) / 1_000_000;
-            log.info("🚀 Consumer average RPS: " + (double) n / ms * 1000);
+            log.info("🚀 Consumer average RPS: " + (int) ((double) n / ms * 1000));
         }
     }
 }
